Rename firewall module to security
