+++ mullvad-daemon/src/tunnel_state_machine/disconnected_state.rs
@@ -2,7 +2,7 @@ use error_chain::ChainedError;
 use futures::sync::mpsc;
 use futures::Stream;
 
-use talpid_core::firewall::NetworkSecurity;
+use talpid_core::security::NetworkSecurity;
 
 use super::{
     ConnectingState, Error, EventConsequence, SharedTunnelStateValues, StateEntryResult,
