+++ talpid-core/src/security/windows/ffi.rs
