+++ talpid-core/src/security/windows/mod.rs
