+++ talpid-core/src/security/windows/system_state.rs
